[xctest] Add test reporters
Not everyone wants to be notified of test results in the same way.

Some people prefer a concise view that fits on a single line, with
dots indicating success and "F" to signify failure: ".....FF..F"

Some prefer a detailed view, containing data on how long each test took
to run, etc.

Objects that conform to the reporter protocol are responsible for
formatting output related to test execution.

Besides providing a way to customize test output, reporters are
essential for another reason: as a way to unit test XCTest itself. Since
XCTest currently prints to stdout, it's difficult to test it without
running it and inspecting its output.

A forthcoming commit will add tests that use a test double as a
reporter, making assertions on what it output by XCTest.