XCTAssertEqualWithAccuracy behavor does not match true XCTest.framework
The accuracy threshold check only works with unsigned numbers, however the distanceTo method we get from Strideable works in terms of offsets which can be negative. We need to take the absolute value of the distance.

This issue was causing tests for https://github.com/apple/swift-corelibs-foundation/pull/37 to fail.