[Tests] Add functional tests and an OS X runner
Adds a test runner that compares annotations in the source code of an
XCTest file to actual output when that source code is compiled and run.
This acts as a regression test suite for the project.

Adds two new Xcode targets:

1. SingleFailingTestCase: An executable that runs `XCTMain()` with a
   failing test case. It is built and installed to the
   Tests/Fixtures/Products/ directory.
2. SwiftXCTestTests: An aggregate target that builds
   SingleFailingTestCase (and, in the future, all test fixtures),
   then runs the test runner, Tests/test.py.

Tests/test.py is a Python unit test script that runs all executables in the
Tests/Fixtures/Products/ directory, then compares their output to source
code annotations in corresponding directories in
Tests/Fixtures/Sources/.