[XCTAssert] Document usage of file and line params
Although it is rare for people to specify these parameters manually, it
is not so rare as to not require documentation. Add a note to the
documentation that explains when they are useful.

Also adopt documentation conventions from the Swift standard library:
use `///` over `/** */`, and only have 80 characters per line.