Refactor messaging further.
Rather than have the _XCTAssertion enum carry the asserted values and
generate the message, use a new _XCTAssertionResult to indicate success
or failure and have it carry the message and generate the full failure
description. This is necessary to be able to use an _XCTAssertion to
generate an assertion name in a context where the values aren't
necessarily available, and also puts the assertion failure messages back
in the assertions themselves.

Reviewed by @mike-ferris-apple.