Allow test methods to throw an error as an unexpected failure
Change the type of the function portion of XCTestCaseProvider.allTests to a
closure that throws, so that test methods are allowed to throw an error. The
only change this requires on clients' part is to adjust their declarations of
allTests.

When a test method does throw an error, treat that as an unexpected failure.

Addresses rdar://problem/23778435.