[build_script] Use built XCTest for Linux tests
Previously, tests on Linux could only be run using the installed
swift-corelibs-xctest. Modify the script to use the built XCTest
instead of the installed one.

In addition, since the script no longer relies on the installed
XCTest, move the `--test` logic such that it can be used even
if `--module-install-path` and `--library-install-path` are not
specified.