[.xcodeproj] Share SwiftXCTest scheme
Currently, running the Swift build script
`swift/utils/build-script --xctest` on OS X fails because it attempts
to use the Linux-only `build_script.py`.

We should make it possible for swift-corelibs-xctest to be built via
the `--xctest` switch on OS X, which means making the project
build-able via the command line.

To make the project build-able via the command line, share the
"SwiftXCTest" scheme. This allows the project to be built via:

```
$ xcodebuild -project SwiftXCTest.xcodeproj -scheme SwiftXCTest
```