[build_script.py] Add "test" subcommand
This "FIXME" describes a problem with the swift-corelibs-xctest build
script when invoked in the context of the greater Swift build script:
https://github.com/apple/swift/blob/6ccf5da2dfc8a74c84b2fd99c32dcdf2255699d7/utils/build-script-impl#L2070-L2072

The problem is that there is currently no way to test an *already
built* XCTest.so. Instead, the XCTest build script re-builds the entire
project, then tests the newly build XCTest.so.

Add a "test" subcommand to build_script.py, which allows users to test
an already built XCTest.so:

```
$ ./build_script.py test --swiftc `which swiftc` /path/to/build/dir
```

Users may still choose to build *and* test, by using the old
build_script.py invocation:

```
$ ./build_script.py --swiftc `which swiftc` --test
```