[build_script.py] Add "install" subcommand
This builds upon the subcommand system added in
https://github.com/apple/swift-corelibs-xctest/pull/57, to add
an install command.

As that pull request describes, because there is no way to test
or install XCTest without also building the project, the current
Swift build script ends up building XCTest three times. This adds
an "install" subcommand that allows an already built XCTest to be
installed.