[build_script.py] Add --foundation-build-dir parameter
https://github.com/apple/swift-corelibs-xctest/pull/43 will add a
dependency between swift-corelibs-xctest and swift-corelibs-foundation.
This will also necessitate the build script to take a
'--foundation-build-dir' parameter. However, the Swift CI does not yet
supply this parameter.

To prevent CI from breaking when #43 is landed, preemptively add the
'--foundation-build-dir' parameter. Before landing #43, the Swift build
script will be modified to pass this parameter to the XCTest build script.