[Python] Add flake8 config
In https://github.com/apple/swift/pull/1153, @practicalswift added a
flake8 config to the Swift project, in order to lint the Python scripts
used by that project. This does the same for XCTest.

This includes all the linting rules active on the main Swift repository.
It also includes tweaks to `build_script.py`, which bring this project in full
compliance with those linting rules.

To lint the Python code in the project:

    $ flake8

To install flake8:

    $ pip install flake8

See https://flake8.readthedocs.org/en/latest/ for details.