[README] Encourage using the Swift build script
New contributors to swift-corelibs-xctest are having some difficulty
knowing where to start:

- https://bugs.swift.org/browse/SR-981
- https://bugs.swift.org/browse/SR-907
- https://bugs.swift.org/browse/SR-906

The problem is that over time this project has begun to take on more
dependencies:

- To build the project, one must use a modern Swift 3 compiler (often
  one more recent than the latest https://swift.org snapshot).
- To build the project, one must have swift-corelibs-foundation checked
  out at a particular path.
- To run the test suite, one must have the LLVM `lit` test runner
  installed at a particular path.

In addition, the build steps differ based on whether the contributor is
using an OS X or Linux development environment.

This commit revamps the README and the Linux `build_script.py` to
require fewer steps for new contributors to get started. Specifically,
it encourages contributors to use the Swift `utils/build-script`. There
are many reasons to do so:

- The XCTest CI uses that script, so contributors will need to make
  sure it passes to get their contributions merged.
- That script ensures that the correct build of `swiftc` and
  swift-corelibs-foundation are used.
- That script works across all platforms, and is invoked in the same
  way on each platform.

In addition, more detailed information on the project has been moved to
a Documentation directory. This resembles other projects in the Swift
family of repositories.