[Tests] Automatically substitute [[@LINE]] numbers
Enhance xctest_checker such that it substitutes the string "[[@LINE]]"
and "[[@LINE+<offset>]]" with the current line number of the source file
when performing checks.

This allows us to add and remove arbitary lines in the functional test
suite without manually updating line numbers in most cases. Line
offsets will still need to be updated when the distance between the
CHECK and the test failure changes.

To test the offsets out, use them in one of the asynchronous tests.