[Tests] Replace %{built_tests_dir} with %T
When I first added functional tests in 589b23d, I struggled against
`lit`'s design: I attempted to prevent test output from being
generated alongside the tests themselves. It wasn't until 8181c16
that I saw the error of my ways: `lit` places test output in the
next to the tests in order to make them easier to debug. That commit
embraced this design, and excluded the test output from source control.

However, we can do better. The XCTest `lit` tests produce test
executables, which the test suite currently attempts to hide in a
tmpdir somewhere. These executables are useful: it's nice to be able to
re-run an executable at will.

Change the location of the where the executables are generated.
Instead of placing them in `%{built_tests_dir}`, place them in an
output directory next to the test itself (a concept built into `lit`
and provided by the `%T` substitution). Since we no longer use
`%{built_tests_dir}`, remove it.