[Performance] Pass correct file/line to methods
Were these two callsites to fail for some reason, they would display the file
and line number for an internal piece of swift-corelibs-xctest source code,
rather than the callsite in the user's test. Thread through the user calliste
file and line to fix the issue.

In practice, these two callsites cannot possibly fail:

- The call to `startMeasuring()` occurs before the user has any chance to
  affect the measurement process, so the meter could not possibly be in an
  invalid state.
- The call to `stopMeasuring()` is only done if measuring is in progress.

Still, if the structure of this code changes in the future, this commit could
prevent issues from popping up.