[build_script] Remove grep since it always passes
`xcodebuild` always prints "export" once it begins building an Xcode project
or workspace, regardless of whether that build eventually passes or fails.
This results in the `grep` statement succeeding, and thus causing the build
script to pass, even when the build fails.

The `grep` statement was originally added to make the build output more human
readable, not change the exit code of any builds. Remove it for now.