[SR-5247] Add the new spellings for the *WithAccuracy assertion functions
This is to bring Corelibs XCTest in line with changes made in the
main XCTest Swift overlay for Xcode 9.

XCTAssertEqualWithAccuracy and XCTAssertNotEqualWithAccuracy are now
deprecated in favor of XCTAssertEqual(_:_:accuracy:file:line:) and
XCTAssertNotEqual(_:_:accuracy:file:line:)

This fixes an inconsistency where XCTAssertEqualWithAccuracy had
accuracy: as a named parameter, but XCTAssertNotEqualWithAccuracy
had it unnamed.