Tests: adjust test for Windows release build
Running the test in release mode on Windows, the block would completely
get optimized out and the resulting time slice was 0 per measurement.
This would result in the relative standard deviation being undefined,
which would render as `nan(ind)%` failing the expectation.  Add a sleep
to ensure that the block has something to execute.