Improve error reporting in edge cases.
When a failure happens after the test suite has stopped (or started),
the failure message included the description of the class XCTestRun or
XCTestCaseRun, which is the default description.

Since those errors are difficult to trace back, instead of logging an
unhelpful source/line location pointing to the inside of XCTest, log the
additional information passed in into the method, to help the user to
locate their error.