Make XCTestExpectation Sendable
This makes the `XCTestExpectation` class hierarchy conform to `Sendable`. This class and its descendants are designed to be used concurrently and are already internally concurrency-safe, so this conformance formally permits them to be used in Swift Concurrency APIs which enforce sendability.

Includes a build-only test which has concurrency warnings enabled to validate the fix.

rdar://95746454