Tests: add vfs overlay for dipsatch
Introduce an additional compile time parameter for swift builds in the
tests to pass along the VFS overlay for dispatch if the build directory
is present.  This ensures that we prefer the just built dispatch over
the installed dispatch which is a prerequisite for permitting building
Foundation et al with a build tree of dispatch when we have an installed
toolchain.