Stop importing CF as a framework
Frameworks don't exist off of Darwin resulting in link failures in this
test. The CF modulemap declares CF as a framework everywhere, which is
wrong, but importing CF from swift-corelibs-foundation is also equally
wrong, which this test does as well, so there's a lot of wrongness.

To get the rebranch working, we need to stop importing CoreFoundation as
a framework in this test, or disable the test on non-darwin platforms.
Given that corelibs-xctest is for non-Darwin platforms, that seems like
the wrong course of action.