Don't use CoreFoundation when building with corelibs-foundation
The use of CoreFoundation while building with corelibs-foundation
causes an unintentional dependency to the CoreFoundation framework due
to the lack of way to distinguish corelibs-foundation's CoreFoundation
from CoreFoundation.framework.

From macOS 13, CoreFoundation started to depend on CoreServicesInternal,
and it causes a transitive dependency from CoreFoundation.framework to
Foundation.framework.

  ┌-> CoreFoundation --[NEW]--> CoreServicesInternal -┐
  └-- Foundation <------------- CFNetwork ------------┘

Thus corelibs-xctest started to depend on Foundation.framework unintentionally
from macOS 13.

This unintentional dependency is problematic because it causes a duplicate
loading of the Objective-C classes of Foundation from the corelibs-foundation
and Foundation.framework even though we explicitly imports corelibs-foundation
by `import SwiftFoundation`.

This patch removes the use of CoreFoundation when building with
corelibs-foundation based on the `USE_FOUNDATION_FRAMEWORK` definition.