Ignore the `--testing-library` argument we are going to start passing from SwiftPM.
This PR adds functionality to XCTest's argument parser that allows it to skip
arguments that the caller may specify but which are uninteresting to XCTest.
Specifically, we are planning to start passing `--testing-library xctest` to the
test runner process in order to disambiguate runs of XCTest from runs of Swift
Testing, and we need to make sure this does not cause XCTest to run incorrectly.

Resolves rdar://129695171.