Test: simplify the test configuration
Remove the compatibility paths in the test infrastructure. Foundation
and dispatch have had evolved over the years. Remove the compatibility
assuming that something within the past few years is used to build. This
is now more reasonable given that the Foundation and dispatch
dependencies have been made optional - this allows building XCTest
against a SDK.

This also reduces the setup for Windows to a very trivial skeleton of
just passing along SDKROOT. This enables the windows build to run the
test suite against the SDKROOT that has been built up which allows us to
swap to the new runtime implementation.